feat: Refactor ResultsExtractor into modular components
Moved sensor, power, SAR, and reporting logic from ResultsExtractor into dedicated classes within the new src/extraction module. This improves modularity and maintainability.
