+++ src/extraction/__init__.py
@@ -1,0 +1,19 @@
+"""Results extraction package.
+
+This package contains specialized modules for extracting different types
+of data from Sim4Life simulation results.
+"""
+
+from .cleaner import Cleaner
+from .power_extractor import PowerExtractor
+from .reporter import Reporter
+from .sar_extractor import SarExtractor
+from .sensor_extractor import SensorExtractor
+
+__all__ = [
+    "Cleaner",
+    "PowerExtractor",
+    "Reporter",
+    "SarExtractor",
+    "SensorExtractor",
+]
